Move poster side component to poster file
