+++ client/src/components/ui/poster/task-top-bar.tsx
